requirements.txt
pandas~=2.3.2
numpy~=2.3.3
scikit-learn~=1.7.2
xgboost~=3.0.5
joblib~=1.5.2
streamlit~=1.50.0
google-generativeai~=0.8.5